fix more dumbdumb mistakes

+++ march_hardware/include/march_hardware/EtherCAT/EthercatSDO.h
@@ -7,9 +7,9 @@
 namespace march
 {
 int sdo_bit8_write(uint16_t slave, uint16_t index, uint8_t sub, uint8_t value);
-int sdo_bit8_read(uint16_t slave, uint16_t index, uint8_t sub, int* val_size, uint32_t* value);
+int sdo_bit8_read(uint16_t slave, uint16_t index, uint8_t sub, int* val_size, uint8_t* value);
 int sdo_bit16_write(uint16_t slave, uint16_t index, uint8_t sub, uint16_t value);
-int sdo_bit16_read(uint16_t slave, uint16_t index, uint8_t sub, int* val_size, uint32_t* value);
+int sdo_bit16_read(uint16_t slave, uint16_t index, uint8_t sub, int* val_size, uint16_t* value);
 int sdo_bit32_write(uint16_t slave, uint16_t index, uint8_t sub, uint32_t value);
 int sdo_bit32_read(uint16_t slave, uint16_t index, uint8_t sub, int* val_size, uint32_t* value);
